machine/raspberry-pi: start moving RPi bootloader functionality out of nixpkgs
The RPi bootloader functionality is deprecated and I don't actually need
most of it. Start to integrate the useful subset of it into this repo.

Also, introduce the nixos-update-firmware command, which allows updating
the firmware on demand, rather than the rather risky approach of
updating along with the system configuration.

Lastly, fix AudioRecorder partition hash computation so each device has
a unique hash.