p-3400: suspend system when idle
The p-3400 doesn't do anything except build software, so it can be
suspended most of the time. Configure the NIC to wake on unicast, and
logind to suspend after 10 seconds idle. This suspend time is very short
because the system has a tendency to wake up immediately a few times
before finally staying asleep. Also needed to disable pings from the
RasPi2 to prevent it waking the system up.