lib/net: fix shadow/coshadow
coerce a (-1) used to produce an address with all bits set, I think due
to sign extension of shifting. Shifts were reimplemented to behave as
unsigned (needed due to overflow checks added in Nix), so this doesn't
work any more. Instead, add a "ones" function that explicitly returns an
address with all bits set. Do this by coercing an IPv6 address (the
largest address) to the desired type.